097_Neural_Rendering_using_Neural_Radiance_Fields_An_Intuition/fwd_code/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
argcomplete=1.12.3=pyhd3eb1b0_0
argon2-cffi=20.1.0=py37h27cfd23_1
async_generator=1.10=py37h28b3542_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
ca-certificates=2021.9.30=h06a4308_1
certifi=2021.10.8=py37h06a4308_0
cffi=1.14.6=py37h400218f_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=pypi_0
dbus=1.13.18=hb2f20db_0
debugpy=1.4.1=py37h295c915_0
decorator=5.1.0=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
dominate=2.6.0=pypi_0
entrypoints=0.3=py37_0
expat=2.4.1=h2531618_2
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
future=0.18.2=pypi_0
fvcore=0.1.5.post20210915=py37
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h5202010_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
imageio=2.9.0=pypi_0
importlib-metadata=4.8.1=py37h06a4308_0
importlib_metadata=4.8.1=hd3eb1b0_0
intel-openmp=2021.3.0=h06a4308_3350
iopath=0.1.9=py37
ipykernel=6.4.1=py37h06a4308_1
ipython=7.27.0=py37hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.0=py37h06a4308_1
jinja2=3.0.1=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter=1.0.0=py37_7
jupyter_client=7.0.1=pyhd3eb1b0_0
jupyter_console=6.4.0=pyhd3eb1b0_0
jupyter_core=4.8.1=py37h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.3.2=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h7f8727e_2
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lpips=0.1.4=pypi_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.0.1=py37h27cfd23_0
matplotlib=3.4.3=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py37h14c3975_1001
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py37h06a4308_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.6.3=pypi_0
ninja=1.10.2=hff7bd54_1
notebook=6.4.3=py37h06a4308_0
numpy=1.21.2=py37h20f2e39_0
numpy-base=1.21.2=py37h79a1101_0
nvidiacub=1.10.0=0
olefile=0.46=py37_0
opencv-python=4.5.4.58=pypi_0
openssl=1.1.1l=h7f8727e_0
packaging=21.0=pyhd3eb1b0_0
pandocfilters=1.4.3=py37h06a4308_1
parso=0.8.2=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=py37h5aabda8_0
pip=21.2.2=py37h06a4308_0
plotly=5.3.1=pypi_0
portalocker=2.3.0=py37h06a4308_0
prometheus_client=0.11.0=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
protobuf=3.19.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pycparser=2.20=py_2
pygments=2.10.0=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.17.3=py37h7b6447c_0
python=3.7.9=h7579374_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.6.0=py3.7_cuda10.1.243_cudnn7.6.3_0
pytorch3d=0.4.0=py37_cu101_pyt160
pywavelets=1.1.1=pypi_0
pyyaml=5.4.1=py37h27cfd23_1
pyzmq=22.2.1=py37h295c915_1
qt=5.9.7=h5867ecd_1
qtconsole=5.1.1=pyhd3eb1b0_0
qtpy=1.10.0=pyhd3eb1b0_0
readline=8.1=h27cfd23_0
scikit-image=0.18.1=pypi_0
scipy=1.7.1=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=58.0.4=py37h06a4308_0
sip=4.19.8=py37hf484d3e_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
tabulate=0.8.9=py37h06a4308_0
tenacity=8.0.1=pypi_0
tensorboardx=2.4=pypi_0
termcolor=1.1.0=py37h06a4308_1
terminado=0.9.4=py37h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
tifffile=2021.10.12=pypi_0
tk=8.6.11=h1ccaba5_0
torchvision=0.7.0=py37_cu101
tornado=6.1=py37h27cfd23_0
tqdm=4.62.2=pyhd3eb1b0_1
traitlets=5.1.0=pyhd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py37_1
wheel=0.37.0=pyhd3eb1b0_1
widgetsnbextension=3.5.1=py37_0
xz=5.2.5=h7b6447c_0
yacs=0.1.6=pyhd3eb1b0_1
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
